Added part 16

+++ rpi_bm/part16/src/printf.c
@@ -26,7 +26,7 @@ static void* stdout_putp;
 static void uli2a(unsigned long int num, unsigned int base, int uc,char * bf)
     {
     int n=0;
-    unsigned long int d=1;
+    unsigned int d=1;
     while (num/d >= base)
         d*=base;
     while (d!=0) {
